refactor: rm list-hashes cmd since no one use it (#247)
